backend/requirements.txt
fastapi
accessify
psycopg2-binary
bestconfig
fastapi-cors
PyJWT
sqlparse